config.py: significantly reduce parallelism
Now that there is botticelli floating around and given that we have no
control of the length of its queue, reduce parallelism of the server.

Should have little effect because load is low in general, but better to
be prudent with respect to this potential issue.

Earlier seven tests were allowed at a time. Now three: one is botticelli
while the others are allowed to run on neubot-server.

This should be more conservative than before, even though, as we know
the queue management is less good than before because now we basically
have a queue for each servent rather than a global queue.

Speaking of this issue, I need to find a way to integrate the queue
management of the two products and I also spent some time thinking at
how this can be done but it seems to me we're too close to release now
to go back to the board and also design for this case. Next release.